pyproject.toml
[tool.poetry]
name = "solis-inverter-datalogger"
version = "0.0.0"
description = "Communication to a Solis inverter using modbus"
authors = ["ivan"]

packages = [
    { include = "src"}
]

[tool.poetry.dependencies]
python = "^3.9"
pyyaml = "6.0"
minimalmodbus = "2.0.1"
python-daemon = "2.3.1"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
pytest = "^7.1.1"
pylint = "^2.13.4"
mypy = "^0.942"
isort = {extras = ["pyproject"], optional = true, version = "*"}
unify = "^0.5"
flake8 = "^4.0.1"
pytest-cov = "^3.0.0"


[tool.poetry.scripts]
run-client="main:main"

[tool.black]
line_length = 140
target-version = ['py37', 'py38', 'py39']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | tests/argo_workflows/compiler/testdata
)/
'''

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
style = "pep440"
bump = true


[build-system]
requires = ["poetry>=1.1.0", "poetry-dynamic-versioning"]
build-backend = "poetry.masonry.api"